first heuristic model

+++ rust/src/strategies/heuristic_player.rs
@@ -3,7 +3,7 @@ use crate::game_components::Action;
 use crate::game::Board;
 
 struct ActionEval {
-  value: i32,
+  value: f64,
   action: Option<Action>,
 }
 impl ActionEval {
@@ -16,28 +16,28 @@ impl ActionEval {
 }
 
 pub struct HeuristicPlayer<'a> {
-  heuristic: &'a dyn Fn(&Board) -> i32,
+  heuristic: &'a dyn Fn(&Board) -> f64,
 }
 impl<'a> HeuristicPlayer<'a> {
-  pub fn new (heuristic: &'a dyn Fn(&Board) -> i32) -> HeuristicPlayer {
+  pub fn new (heuristic: &'a dyn Fn(&Board) -> f64) -> HeuristicPlayer {
     HeuristicPlayer {
       heuristic
     }
   }
-  pub fn boxed_new (heuristic: &'a dyn Fn(&Board) -> i32) -> Box<HeuristicPlayer> {
+  pub fn boxed_new (heuristic: &'a dyn Fn(&Board) -> f64) -> Box<HeuristicPlayer> {
     Box::new(Self::new(heuristic))
   }
 
   fn search_action (&self, depth:i32, board:&Board) -> ActionEval {
     if board.is_ended() {
-      return ActionEval { value:board.ended_value(board.now_player), action:None };
+      return ActionEval { value:board.ended_value(board.now_player) as f64, action:None };
     }
     if depth == 0 {
       return ActionEval { value:(self.heuristic)(board), action:None };
     }
 
     let mut eval = ActionEval {
-      value: i32::MIN,
+      value: f64::MIN,
       action: None,
     };
